Update bootstrap/components js
