+++ resources/assets/js/components.js
@@ -1,3 +1,20 @@
+window.Vue = require('vue');
+import BootstrapVue from 'bootstrap-vue'
+Vue.use(BootstrapVue);
+
+window.InfiniteScroll = require('infinite-scroll');
+window.filesize = require('filesize');
+window.axios = require('axios');
+window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
+import swal from 'sweetalert';
+
+let token = document.head.querySelector('meta[name="csrf-token"]');
+if (token) {
+    window.axios.defaults.headers.common['X-CSRF-TOKEN'] = token.content;
+} else {
+    console.error('CSRF token not found: https://laravel.com/docs/csrf#csrf-x-csrf-token');
+}
+
 require('./components/localstorage');
 require('./components/likebutton');
 require('./components/commentform');
